VS config api
